chore(node): reorganize test macros
- exec macros now in `test/unit/exec-macros.js`
- CLI macros now in `test/e2e/cli-macros.js`
- created new `testExecForJSON` macro; add test
- rename `test/unit/policy/macros.js` => `test/unit/policy/policy-macros.js`
- rename `execution.spec.js` => `exec.spec.js`
- create constant for JSON snapshot entry point file (`/index.js`)
- fixed absolute path creeping in to snapshot markdown
- better setup/teardown in CLI test
- rebuild & cleanup snapshots
- add a `test:update-snapshots` script
