+++ packages/node/test/e2e/snapshots/cli.spec.js.md
@@ -6,7 +6,7 @@ Generated by [AVA](https://avajs.dev).
 
 ## "--help" prints help
 
-> output of command "/opt/homebrew/Cellar/nvm/0.40.1/versions/node/v22.11.0/bin/node /Users/boneskull/projects/lavamoat/lavamoat/packages/node/src/cli.js --help" does not match expected snapshot
+> output of command "node ../../../src/cli.js --help" does not match expected snapshot
 
     {
       code: undefined,
@@ -30,7 +30,7 @@ Generated by [AVA](https://avajs.dev).
                                                [string] [default: (current directory)]␊
       ␊
       Behavior Options:␊
-            --dev  Include development dependencies          [boolean] [default: true]␊
+            --dev  Include development dependencies                          [boolean]␊
       ␊
       Positionals:␊
         entrypoint  Path to the application entry point; relative to --root   [string]␊
@@ -48,7 +48,7 @@ Generated by [AVA](https://avajs.dev).
 
 ## run - "run --help" prints help
 
-> output of command "/opt/homebrew/Cellar/nvm/0.40.1/versions/node/v22.11.0/bin/node /Users/boneskull/projects/lavamoat/lavamoat/packages/node/src/cli.js run --help" does not match expected snapshot
+> output of command "node ../../../src/cli.js run --help" does not match expected snapshot
 
     {
       code: undefined,
@@ -62,7 +62,7 @@ Generated by [AVA](https://avajs.dev).
                                                                    [string] [required]␊
       ␊
       Behavior Options:␊
-            --dev  Include development dependencies          [boolean] [default: true]␊
+            --dev  Include development dependencies                          [boolean]␊
       ␊
       Path Options:␊
         -p, --policy           Filepath to a policy file␊
@@ -81,7 +81,7 @@ Generated by [AVA](https://avajs.dev).
 
 ## generate - "generate --help" prints help
 
-> output of command "/opt/homebrew/Cellar/nvm/0.40.1/versions/node/v22.11.0/bin/node /Users/boneskull/projects/lavamoat/lavamoat/packages/node/src/cli.js generate --help" does not match expected snapshot
+> output of command "node ../../../src/cli.js generate --help" does not match expected snapshot
 
     {
       code: undefined,
@@ -91,7 +91,7 @@ Generated by [AVA](https://avajs.dev).
       Generate a policy␊
       ␊
       Behavior Options:␊
-            --dev    Include development dependencies        [boolean] [default: true]␊
+            --dev    Include development dependencies                        [boolean]␊
             --debug  Additionally write a debug policy                       [boolean]␊
             --write  Write policy file(s) to disk; if false, print to stdout␊
                                                              [boolean] [default: true]␊
