+++ packages/node/test/unit/json-fixture-util.js
@@ -126,3 +126,8 @@ loadJSONFixture.snapshotCache = new Map()
  * The trailing slash is load-bearing.
  */
 export const JSON_FIXTURE_DIR_URL = new URL('./json-fixture/', import.meta.url)
+
+/**
+ * The entry point for all JSON fixtures
+ */
+export const JSON_FIXTURE_ENTRY_POINT = '/index.js'
