+++ packages/node/test/unit/policy/policy-macros.js
@@ -1,6 +1,10 @@
 import { generatePolicy } from '../../../src/policy-gen/generate.js'
 import { isPolicy } from '../../../src/policy-util.js'
-import { JSON_FIXTURE_DIR_URL, loadJSONFixture } from '../json-fixture-util.js'
+import {
+  JSON_FIXTURE_DIR_URL,
+  JSON_FIXTURE_ENTRY_POINT,
+  loadJSONFixture,
+} from '../json-fixture-util.js'
 
 /**
  * @import {ValueOf} from 'type-fest'
@@ -171,7 +175,7 @@ export function createGeneratePolicyMacros(test) {
         new URL(fixtureFilename, JSON_FIXTURE_DIR_URL)
       )
 
-      const actualPolicy = await generatePolicy('/index.js', {
+      const actualPolicy = await generatePolicy(JSON_FIXTURE_ENTRY_POINT, {
         ...options,
         readPowers,
       })
