+++ packages/node/test/unit/policy/snapshots/build-module-records.spec.js.md
@@ -9,9 +9,3 @@ Generated by [AVA](https://avajs.dev).
 > Snapshot 1
 
     '[{"file":"/node_modules/app/main.js","importMap":{"app":"/node_modules/app/main.js","avery":"/node_modules/avery/avery.js","brooke":"/node_modules/brooke/brooke.js","cjs":"/node_modules/cjs/index.js","clarke":"/node_modules/clarke/index.js","danny":"/node_modules/danny/src/index.js"},"packageName":"app","specifier":"/node_modules/app/main.js","type":"js"},{"file":"/node_modules/avery/avery.js","importMap":{},"packageName":"avery","specifier":"/node_modules/avery/avery.js","type":"js"},{"file":"/node_modules/brooke/brooke.js","importMap":{"clarke":"/node_modules/clarke/index.js"},"packageName":"brooke","specifier":"/node_modules/brooke/brooke.js","type":"js"},{"file":"/node_modules/cjs/index.js","importMap":{"clarke":"/node_modules/cjs/node_modules/clarke/index.js"},"packageName":"cjs","specifier":"/node_modules/cjs/index.js","type":"js"},{"file":"/node_modules/clarke/index.js","importMap":{},"packageName":"clarke","specifier":"/node_modules/clarke/index.js","type":"js"},{"file":"/node_modules/cjs/node_modules/clarke/index.js","importMap":{},"packageName":"clarke","specifier":"/node_modules/cjs/node_modules/clarke/index.js","type":"js"},{"file":"/node_modules/danny/src/index.js","importMap":{"danny/src/danny.js":"/node_modules/danny/src/danny.js"},"packageName":"danny","specifier":"/node_modules/danny/src/index.js","type":"js"},{"file":"/node_modules/danny/src/danny.js","importMap":{},"packageName":"danny","specifier":"/node_modules/danny/src/danny.js","type":"js"},{"file":"/node_modules/edwin/index.js","importMap":{},"packageName":"edwin","specifier":"/node_modules/edwin/index.js","type":"js"},{"file":"/index.js","importMap":{"app":"/node_modules/app/main.js"},"packageName":"$root$","specifier":"/index.js","type":"js"}]'
-
-## buildModuleRecords() - result is deterministic
-
-> Snapshot 1
-
-    '[{"file":"/node_modules/app/main.js","importMap":{"app":"/node_modules/app/main.js","avery":"/node_modules/avery/avery.js","brooke":"/node_modules/brooke/brooke.js","cjs":"/node_modules/cjs/index.js","clarke":"/node_modules/clarke/index.js","danny":"/node_modules/danny/src/index.js"},"packageName":"app","specifier":"/node_modules/app/main.js","type":"js"},{"file":"/node_modules/avery/avery.js","importMap":{},"packageName":"avery","specifier":"/node_modules/avery/avery.js","type":"js"},{"file":"/node_modules/brooke/brooke.js","importMap":{"clarke":"/node_modules/clarke/index.js"},"packageName":"brooke","specifier":"/node_modules/brooke/brooke.js","type":"js"},{"file":"/node_modules/cjs/index.js","importMap":{"clarke":"/node_modules/cjs/node_modules/clarke/index.js"},"packageName":"cjs","specifier":"/node_modules/cjs/index.js","type":"js"},{"file":"/node_modules/clarke/index.js","importMap":{},"packageName":"clarke","specifier":"/node_modules/clarke/index.js","type":"js"},{"file":"/node_modules/cjs/node_modules/clarke/index.js","importMap":{},"packageName":"clarke","specifier":"/node_modules/cjs/node_modules/clarke/index.js","type":"js"},{"file":"/node_modules/danny/src/index.js","importMap":{"danny/src/danny.js":"/node_modules/danny/src/danny.js"},"packageName":"danny","specifier":"/node_modules/danny/src/index.js","type":"js"},{"file":"/node_modules/danny/src/danny.js","importMap":{},"packageName":"danny","specifier":"/node_modules/danny/src/danny.js","type":"js"},{"file":"/node_modules/edwin/index.js","importMap":{},"packageName":"edwin","specifier":"/node_modules/edwin/index.js","type":"js"},{"file":"/index.js","importMap":{"app":"/node_modules/app/main.js"},"packageName":"$root$","specifier":"/index.js","type":"js"}]'
