--- packages/node/test/unit/policy/snapshots/generate-policy.spec.js.md
@@ -1,139 +1,0 @@
-# Snapshot report for `test/policy-gen/generate-policy.spec.js`
-
-The actual snapshot is saved in `generate-policy.spec.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## Node.js builtins (builtins.json)
-
-> Snapshot 1
-
-    {
-      resources: {
-        a: {
-          packages: {
-            events: true,
-          },
-        },
-        b: {
-          builtin: {
-            'events.EventEmitter': true,
-          },
-        },
-        events: {
-          globals: {
-            console: true,
-          },
-        },
-      },
-    }
-
-## kitchen sink (kitchen-sink.json)
-
-> Snapshot 1
-
-    {
-      resources: {
-        app: {
-          globals: {
-            globalProperty: true,
-          },
-          packages: {
-            avery: true,
-            brooke: true,
-            cjs: true,
-            clarke: true,
-            danny: true,
-          },
-        },
-        brooke: {
-          packages: {
-            clarke: true,
-          },
-        },
-        cjs: {
-          packages: {
-            clarke: true,
-          },
-        },
-      },
-    }
-
-## native-like module (phony-native.json)
-
-> Snapshot 1
-
-    {
-      resources: {
-        hello: {
-          globals: {
-            'console.log': true,
-          },
-          native: true,
-        },
-      },
-    }
-
-## native module (native.json)
-
-> Snapshot 1
-
-    {
-      resources: {
-        hello_world: {
-          globals: {
-            __dirname: true,
-          },
-          packages: {
-            'node-gyp-build': true,
-          },
-        },
-        'node-gyp-build': {
-          builtin: {
-            'fs.existsSync': true,
-            'fs.readdirSync': true,
-            'os.arch': true,
-            'os.platform': true,
-            'path.dirname': true,
-            'path.join': true,
-            'path.resolve': true,
-          },
-          globals: {
-            __non_webpack_require__: true,
-            __webpack_require__: true,
-            process: true,
-          },
-        },
-      },
-    }
-
-## native module w/ dynamic requires (dynamic.json)
-
-> Snapshot 1
-
-    {
-      resources: {},
-    }
-
-## ignoring unused dependencies (unused-dependency.json)
-
-> Snapshot 1
-
-    {
-      resources: {},
-    }
-
-## policy override merging (unused-dependency-override.json)
-
-> Snapshot 1
-
-    {
-      resources: {
-        blurmph: {
-          builtin: {
-            'node:fs': true,
-            'node:path': true,
-          },
-        },
-      },
-    }
